requirements.txt
appdirs==1.4.4
audioread==3.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
decorator==5.1.1
ffmpeg-python==0.2.0
filelock==3.12.0
future==0.18.3
idna==3.4
imap-tools==1.0.0
importlib-metadata==6.6.0
Jinja2==3.1.2
joblib==1.2.0
lazy_loader==0.2
librosa==0.10.0.post2
llvmlite==0.40.0
MarkupSafe==2.1.2
more-itertools==9.1.0
mpmath==1.3.0
msgpack==1.0.5
networkx==3.1
numba==0.57.0
numpy==1.24.3
openai-whisper==20230314
packaging==23.1
pooch==1.6.0
pycparser==2.21
regex==2023.5.5
requests==2.30.0
scikit-learn==1.2.2
scipy==1.10.1
soundfile==0.12.1
soxr==0.3.5
sympy==1.11.1
threadpoolctl==3.1.0
tiktoken==0.3.1
torch==2.0.0
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==2.0.2
zipp==3.15.0
